limited table height and fixed open modal
