+++ src/components/Table/TableRow.tsx
@@ -1,72 +1,24 @@
-import React, { useState, useEffect } from 'react';
+import React from 'react';
 import { useLocation } from 'react-router-dom';
 import resources from '../../resources/resources.ts';
 import { useDispatch, useSelector } from 'react-redux';
 import { AppDispatch, RootState } from '../../redux/store.ts';
 import { BsFillPencilFill, BsFillTrashFill, BsCopy, BsSave, BsX } from 'react-icons/bs';
-import useCollection from '../../hooks/useCollection.ts';
-import * as thunks from '../../redux/tables/tableThunks.ts';
+import useCollection from '../../hooks/useCollection.tsx';
 import * as constants from '../../resources/constants.ts';
 import { setDocument } from '../../redux/tables/tableSlice.ts';
+import useHandlers from '../../hooks/useHandlers.tsx';
 
 function TableRow({ item }) {
     const dispatch = useDispatch<AppDispatch>();
-    const currLocation = useLocation().pathname;
     const currType = useCollection()?.type;
-    const [changedItem, setItem] = useState({ ...item });
     const currDocument = useSelector((state: RootState) => state.collection.currDocument)
-    
-    useEffect(() => {
-        if (currDocument === changedItem._id) {
-            const cloudinaryWidget = window.cloudinary.createUploadWidget(
-                {
-                    cloudName: 'dhavutxxt',
-                    uploadPreset: 'epicure-admin'
-                },
-                (error, result) => {
-                    if (!error && result && result.event === 'success') {
-                        const newImageUrl = result.info.secure_url;
-                        setItem({ ...changedItem, image: newImageUrl });
-                    }
-                }
-            );
 
-            document.getElementById(`image_${changedItem._id}`)?.addEventListener('click', () => {
-                cloudinaryWidget.open();
-            });
-
-            return () => {
-                document.getElementById(`image_${changedItem._id}`)?.removeEventListener('click', () => {
-                    cloudinaryWidget.open();
-                });
-            };
-        }
-    }, [currDocument, changedItem]);
+    const { changedItem, handleCopy, handleSave, handleDelete, handleChange, resetChanges } = useHandlers(item, currType);
 
     if (!currType) {
         return alert(constants.TABLE_CONSTANTS.ROUTE_ERROR);
     }
-
-    const handleCopy = async (toCopy: typeof currType) => {
-        const { _id, ...itemWithoutId } = toCopy;
-        dispatch(thunks.addData({ route: currLocation, item: itemWithoutId }));
-    };
-
-    const handleSave = async (toSave: typeof currType) => {
-        const { _id, ...itemWithoutId } = toSave;
-        dispatch(thunks.updateData({ route: `${currLocation}/${toSave._id}`, item: itemWithoutId }));
-        dispatch(setDocument(undefined));
-    };
-
-    const handleDelete = async (toDelete: typeof currType) => {
-        if (window.confirm(resources.DELETE_CONFIRMATION)) {
-            dispatch(thunks.deleteData({ route: `${currLocation}/${toDelete._id}`, item: toDelete }));
-        }
-    };
-
-    const handleChange = (e, key) => {
-        setItem({ ...changedItem, [key]: e.target.value });
-    };
 
     return (
         <tr key={item._id}>
@@ -132,9 +84,7 @@ function TableRow({ item }) {
                     <td>
                         <span className="actions">
                             <BsSave onClick={() => handleSave(changedItem)} />
-                            <BsX onClick={() => {setItem({...item})
-                                dispatch(setDocument(undefined));
-                            }} />
+                            <BsX onClick={() => resetChanges} />
                         </span>
                     </td>
                 </>
