+++ src/hooks/useHandlers.tsx
@@ -1,0 +1,75 @@
+import react, {useEffect, useState} from 'react';
+import { useDispatch, useSelector } from 'react-redux';
+import { useLocation } from 'react-router';
+import * as thunks from '../redux/tables/tableThunks.ts';
+import { setDocument } from '../redux/tables/tableSlice.ts';
+import { ICollection } from '../types/collectionType.ts';
+import { AppDispatch, RootState } from '../redux/store.ts';
+
+function useHandlers(item: ICollection, currType) {
+    const dispatch = useDispatch<AppDispatch>();
+    const currLocation = useLocation().pathname;
+    const currDocument = useSelector((state: RootState) => state.collection.currDocument);
+    const [changedItem, setItem] = useState({ ...item });
+
+    useEffect(() => {
+        if (currDocument === changedItem._id) {
+            const cloudinaryWidget = window.cloudinary.createUploadWidget(
+                {
+                    cloudName: 'dhavutxxt',
+                    uploadPreset: 'epicure-admin'
+                },
+                (error, result) => {
+                    if (!error && result && result.event === 'success') {
+                        const newImageUrl = result.info.secure_url;
+                        setItem({ ...changedItem, image: newImageUrl });
+                    }
+                }
+            );
+
+            const imageClickHandler = () => cloudinaryWidget.open();
+            document.getElementById(`image_${changedItem._id}`)?.addEventListener('click', imageClickHandler);
+
+            return () => {
+                document.getElementById(`image_${changedItem._id}`)?.removeEventListener('click', imageClickHandler);
+            };
+        }
+    }, [currDocument, changedItem]);
+
+    const handleCopy = async (toCopy) => {
+        const { _id, ...itemWithoutId } = toCopy;
+        dispatch(thunks.addData({ route: currLocation, item: itemWithoutId }));
+    };
+
+    const handleSave = async (toSave) => {
+        const { _id, ...itemWithoutId } = toSave;
+        dispatch(thunks.updateData({ route: `${currLocation}/${toSave._id}`, item: itemWithoutId }));
+        dispatch(setDocument(undefined));
+    };
+
+    const handleDelete = async (toDelete) => {
+        if (window.confirm('Are you sure you want to delete this item?')) {
+            dispatch(thunks.deleteData({ route: `${currLocation}/${toDelete._id}`, item: toDelete }));
+        }
+    };
+
+    const handleChange = (e, key) => {
+        setItem({ ...changedItem, [key]: e.target.value });
+    };
+
+    const resetChanges = () => {
+        setItem({ ...item });
+        dispatch(setDocument(undefined));
+    };
+
+    return {
+        changedItem,
+        handleCopy,
+        handleSave,
+        handleDelete,
+        handleChange,
+        resetChanges
+    };
+};
+
+export default useHandlers;
