+++ src/pages/dashboard.scss
@@ -2,4 +2,6 @@
     display: flex;
     flex-direction: column;
     width: 100%;
+    height: 100vh;
+    overflow-y: hidden;
 }
