+++ src/redux/tables/tableSlice.ts
@@ -37,6 +37,18 @@ const tableSlice = createSlice({
         state.status = constants.STATUS_CODE.IDLE;
         state.collectionData = action.payload;
       })
+      // .addCase(thunks.fetchData.fulfilled, (state, action) => {
+      //   state.status = constants.STATUS_CODE.IDLE;
+      //   state.collectionData = action.payload.map((item)=>{
+      //     const obj = {}
+      //     Object.keys(item).forEach((keyItem)=>{
+      //       obj[keyItem]=  isObj? item[key].name :
+
+      //     })
+      //     return
+      //   });
+
+      // })
 
       .addCase(thunks.updateData.fulfilled, (state, action) => {
         state.status = constants.STATUS_CODE.IDLE;
